Add email testing feature and update welcome email
This commit introduces a new feature that allows admins to test the sending of welcome emails to new users. A new 'Tests' section has been added in the admin navigation for this purpose. The code also includes error handling for failed email sends.

In addition, changes have been made to the content of the welcome email sent to new users. The invitation to join our community has been removed from the message. Also, if a username is not available, the user's email will be used in the greeting instead.